% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_grand_mean.R
\name{center_grand_mean}
\alias{center_grand_mean}
\title{Center with respect to grand mean}
\usage{
center_grand_mean(data, cols)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or quos(). column(s) that need to be centered}
}
\value{
return a dataframe with the columns centered (replace existing columns)
}
\description{
Center all columns with respect to the grand mean
}
\examples{
center_group_mean(data = df1, cols = quos('IV1:IV5'))

}
