% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_score.R
\name{composite_score}
\alias{composite_score}
\title{Composite column}
\usage{
composite_score(data, cols, composite_col_name = "composited_column")
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or quos(). column(s) that need to be composite}

\item{composite_col_name}{character. default as 'composited_column'. the column name of the composite column}
}
\value{
return a dataframe with a new column with the composite score
}
\description{
The function will perform a rowise aggregation which then divided by the total number of columns. It ignores rows with any NA value.
}
\examples{

}
