% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Recode}
\usage{
cor_test(data, cols)
}
\arguments{
\item{data}{a dataframe}

\item{cols}{vector or quos(). column(s) that need to be recoded}
}
\value{

}
\description{
Recode
}
\examples{

}
