% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listwise_deletion.R
\name{listwise_deletion}
\alias{listwise_deletion}
\title{Listwise deletion}
\usage{
listwise_deletion(data, cols)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or quos(). column(s) that need to be centered}
}
\value{
return a dataframe with listwise deletion
}
\description{
Perform listwise deletion (the entire rows is disregarded if the row has one NA value)
}
\examples{


}
