% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{RMarkdown Render}
\usage{
render_all(dir, render_exist = T)
}
\arguments{
\item{dir}{directory path}
}
\value{

}
\description{
Center all columns with respect to the grand mean
}
\examples{

}
