% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_non_missing_values.R
\name{summarize_non_missing_values}
\alias{summarize_non_missing_values}
\title{Count number of non-missing / non-NA values}
\usage{
summarize_non_missing_values(data, cols, group = NULL)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or quos(). column(s) that need to be centered}

\item{group}{character. pass to group_by}
}
\value{
return a dataframe with the number of non-NA values
}
\description{
Center all columns with respect to the grand mean
}
\examples{
summarize_non_missing_values(data = df1, cols = quos(Q1:Q10), group = 'Country')

}
