% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_group_mean.R
\name{center_group_mean}
\alias{center_group_mean}
\title{Center with respect to group mean}
\usage{
center_group_mean(data, cols, group)
}
\arguments{
\item{cols}{vector or quos(). column(s) that need to be centered}

\item{group}{the grouping variable. If you need to pass multiple group variables, try to use quos(). Passing multiple group variables is not tested.}

\item{dataframe}{dataframe}
}
\value{
return a dataframe with the columns centered (replace existing columns)
}
\description{
Center all columns with respect to the group mean
}
\examples{
center_group_mean(data = df1, cols = quos('IV1:IV5'), group = 'Country')
}
