% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_item.R
\name{recode_item}
\alias{recode_item}
\title{Recode}
\usage{
recode_item(data, cols, code_from, code_to)
}
\arguments{
\item{data}{a dataframe}

\item{cols}{vector or quos(). column(s) that need to be recoded}

\item{code_from}{vector. the order must match with vector for code_to}

\item{code_to}{vector. the order must match with vector for code_from}
}
\value{

}
\description{
Recode
}
\examples{

}
